+++ benchmark/include/tpcc/workload.h
@@ -36,7 +36,7 @@ struct TransactionArgs {
 // 2.4.1
 template <class Random>
 TransactionArgs BuildNewOrderArgs(Random *const generator, const int32_t w_id, const uint32_t num_warehouses) {
-  TERRIER_ASSERT(w_id >= 1 && w_id <= num_warehouses, "Invalid w_id.");
+  TERRIER_ASSERT(w_id >= 1 && static_cast<uint32_t>(w_id) <= num_warehouses, "Invalid w_id.");
   TransactionArgs args;
   args.type = TransactionType::NewOrder;
   args.w_id = w_id;
@@ -74,7 +74,7 @@ TransactionArgs BuildNewOrderArgs(Random *const generator, const int32_t w_id, c
 // 2.5.1
 template <class Random>
 TransactionArgs BuildPaymentArgs(Random *const generator, const int32_t w_id, const uint32_t num_warehouses) {
-  TERRIER_ASSERT(w_id >= 1 && w_id <= num_warehouses, "Invalid w_id.");
+  TERRIER_ASSERT(w_id >= 1 && static_cast<uint32_t>(w_id) <= num_warehouses, "Invalid w_id.");
   TransactionArgs args;
   args.type = TransactionType::Payment;
   args.w_id = w_id;
